RF AT_12.002.002|[Teacher] Verify the announcement is not created when the empty form is submitted
