+++ pytest.ini
@@ -1,6 +1,7 @@
 [pytest]
 
 addopts = -s -vv -m "not slow" --alluredir=allure-results --video on --screenshot on
+ --ignore=obsolete_tests
 
 #addopts = -s -vv -m "not slow" --alluredir=allure-results --video on --screenshot on --headed --slowmo=3000
 #for headed UI
