+++ tests/test_register.py
@@ -8,17 +8,16 @@ STUDENT_SUCCESS_REGISTRSTION_MESSAGE = "Вы успешно зарегистри
 text = "Вы успешно зарегистрировались, а так же получаете бесплатный премиум на 3 дня!"
 
 
-@pytest.mark.skip(reason="Тест временно отключен после обновления 09.01.2025")
+@pytest.mark.skip(reason="Функциональность изменена, тест устарел. 09.01.2025")
 def test_register_as_tutor(header, register):
     header.visit()
     header.click_registration_button()
     register.header_should_contain_text("Регистрация")
     register.registration_new_user("tutor")
-    # header.create_listing_button_should_be_visible()
 
 
 # @pytest.mark.slow
-@pytest.mark.skip(reason="Тест временно отключен после обновления 09.01.2025")
+@pytest.mark.skip(reason="Функциональность изменена, тест устарел. 09.01.2025")
 def test_register_as_student(header, register):
     header.visit()
     header.click_registration_button()
@@ -26,7 +25,7 @@ def test_register_as_student(header, register):
     register.registration_new_user("student")
 
 
-@pytest.mark.skip(reason="Тест временно отключен после обновления 09.01.2025")
+@pytest.mark.skip(reason="Функциональность изменена, тест устарел. 09.01.2025")
 def test_register_as_student_verify_success_message_text(
     homepage, register, find_tutor
 ):
@@ -52,7 +51,7 @@ def test_become_a_teacher_from_student_page(
     register.register_page_contains_register_btn()
 
 
-@pytest.mark.skip(reason="Тест временно отключен после обновления 09.01.2025")
+@pytest.mark.skip(reason="Функциональность изменена, тест устарел. 09.01.2025")
 def test_verify_successful_message_after_register_as_tutor(
     homepage, register, find_tutor
 ):
@@ -62,15 +61,3 @@ def test_verify_successful_message_after_register_as_tutor(
     register.header_should_contain_text("Регистрация")
     register.registration_new_user("tutor")
     find_tutor.check_message_of_registration(text)
-
-
-@pytest.mark.skip(reason="Тест временно отключен после обновления 09.01.2025")
-def test_check_placeholders_on_the_register_page(header, register):
-    header.visit()
-    header.click_registration_button()
-    register.check_username_placeholder_visibility()
-    register.check_username_placeholder_text()
-    register.check_password1_placeholder_visibility()
-    register.check_password1_placeholder_text()
-    register.check_password2_placeholder_visibility()
-    register.check_password2_placeholder_text()
